Refactored the code, moving all factories into a Factory package, and creating a PreferenceBookFactory which is called from PreferenceBookCSVReader
