+++ src/main/java/UseCases/Factories/GroupFactory.java
@@ -1,4 +1,4 @@
-package main.java.UseCases;
+package main.java.UseCases.Factories;
 
 import main.java.Entities.Group;
 import java.util.ArrayList;
